💄 [Design] 컴포넌트로 분리 및 사이즈 조정 #86
