+++ client/src/components/User/UserThumnailArea/UserThumnailArea.tsx
@@ -1,0 +1,47 @@
+import {
+  S_UserThumnailArea,
+  S_UserProfileIamge,
+  S_UserFollowContainer,
+  S_UserFollowCount,
+  S_UserFollow,
+} from './UserThumnailArea.styles';
+import Button from '../../common/Button/Button';
+
+import { DataType } from '../../../pages/User/User';
+
+interface IProps {
+  userdata: DataType;
+}
+
+function UserThumnailArea({ userdata }: IProps) {
+  return (
+    <S_UserThumnailArea>
+      <S_UserProfileIamge
+        alt="user profile"
+        src={userdata.data[0].profileImageUrl}
+      />
+      <S_UserFollowContainer>
+        <div>
+          <S_UserFollowCount>
+            {userdata.data[0].followerCount}
+          </S_UserFollowCount>
+          <S_UserFollow>Follower</S_UserFollow>
+        </div>
+        <div>
+          <S_UserFollowCount>
+            {userdata.data[0].followingCount}
+          </S_UserFollowCount>
+          <S_UserFollow>Following</S_UserFollow>
+        </div>
+      </S_UserFollowContainer>
+      <Button variant="point" shape="default" size="large">
+        Upload Image
+      </Button>
+      <Button variant="point" shape="default" size="XLarge">
+        Remove Image
+      </Button>
+    </S_UserThumnailArea>
+  );
+}
+
+export default UserThumnailArea;
